I've addressed the `ImportError` for `get_tasks_by_project_id_ordered_by_sequence`.
I added the missing database function `get_tasks_by_project_id_ordered_by_sequence` to `db/cruds/tasks_crud.py` and ensured it is correctly exported through `db/__init__.py`.

This function is required by `ProductionOrderDetailDialog` and `EditProductionOrderDialog` (within the `project_management` package) and was causing an `ImportError` after the initial refactoring of `projectManagement.py`.

Testing this fix was hindered by persistent environment errors, but the structural changes to resolve the specific `ImportError` have been applied.
