+++ project_management/README.md
@@ -1,0 +1,13 @@
+# Project Management Tools
+
+This directory contains tools and scripts for managing the project.
+
+## Available Tools:
+
+* **Task Tracker:** (Coming Soon) A tool to track project tasks and their status.
+* **Reporting Tool:** (Coming Soon) A tool to generate project reports.
+* **Resource Allocator:** (Coming Soon) A tool to manage and allocate project resources.
+
+## Contribution Guidelines:
+
+Contributions to these tools are welcome. Please follow the project's general contribution guidelines.
